+++ libs/cardano-ledger-core/CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## 1.21.0.0
 
+* Add `directDepositsMissingAccounts` and `applyDirectDeposits`
+* Add `subtractCompactCoin`
 * Add `withdrawalsThatExceedAccountBalance` and `applyWithdrawals`
 * Change `ChainCode` type to use `ByteArray` instead of `ByteString`
 * Change `bwAttributes` field to use `ByteArray` instead of `ByteString`